requirements.txt
python-telegram-bot
instaloader
yt-dlp
gallery-dl